+++ Lolighter/Methods/Sliders.cs
@@ -13,6 +13,8 @@ namespace Lolighter.Methods
             List<BeatmapNote> toRemove = new List<BeatmapNote>();
             BeatmapNote n = new BeatmapNote(0, 0, 0, 0, 0);
             BeatmapNote lastNote = new BeatmapNote(0, 0, 0, 0, 0);
+            float secondNote = Options.Modifier.SliderPrecision > 0.0f ? 1 / Options.Modifier.SliderPrecision : 0.0f;
+            float thirdNote = secondNote * 2;
             bool found = false;
 
             for (int i = noteTemp.Count() - 1; i > -1; i--) //For each note in reverse-order
@@ -73,41 +75,41 @@ namespace Lolighter.Methods
                                 switch (n.CutDirection)
                                 {
                                     case CutDirection.Up:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Middle, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Middle, n.Type, CutDirection.Any));
                                         if (n.Type == NoteType.Red)
                                         {
-                                            newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.Left, Layer.Top, n.Type, CutDirection.Any));
+                                            newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.Left, Layer.Top, n.Type, CutDirection.Any));
                                         }
                                         else
                                         {
-                                            newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.MiddleLeft, Layer.Top, n.Type, CutDirection.Any));
+                                            newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.MiddleLeft, Layer.Top, n.Type, CutDirection.Any));
                                         }
                                         break;
                                     case CutDirection.Down:
                                         n.LineLayer = Layer.Top;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Middle, n.Type, CutDirection.Any));
-                                        newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.Left, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Middle, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.Left, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.Left:
                                         n.LineIndex = Index.MiddleLeft;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.Right:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleLeft, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleLeft, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.UpLeft:
                                         n.LineIndex = 1;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Middle, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Middle, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.UpRight:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleRight, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleRight, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.DownLeft:
                                         n.LineIndex = Index.MiddleLeft;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.DownRight:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleLeft, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleLeft, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                 }
                                 break;
@@ -115,29 +117,29 @@ namespace Lolighter.Methods
                                 switch (n.CutDirection)
                                 {
                                     case CutDirection.Up:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.MiddleLeft, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.MiddleLeft, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.Left:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.Right:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleRight, Layer.Bottom, n.Type, CutDirection.Any));
-                                        newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.Right, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleRight, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.Right, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.UpLeft:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Middle, n.Type, CutDirection.Any));
-                                        newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.Left, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Middle, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.Left, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.UpRight:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.DownLeft:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.DownRight:
                                         n.LineLayer = Layer.Middle;
                                         n.LineIndex = Index.Left;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleLeft, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleLeft, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                 }
                                 break;
@@ -145,29 +147,29 @@ namespace Lolighter.Methods
                                 switch (n.CutDirection)
                                 {
                                     case CutDirection.Up:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.MiddleRight, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.MiddleRight, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.Left:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleLeft, Layer.Bottom, n.Type, CutDirection.Any));
-                                        newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.Left, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleLeft, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.Left, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.Right:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.UpLeft:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.UpRight:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Middle, n.Type, CutDirection.Any));
-                                        newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.Right, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Middle, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.Right, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.DownLeft:
                                         n.LineLayer = Layer.Middle;
                                         n.LineIndex = Index.Right;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleRight, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleRight, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.DownRight:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                 }
                                 break;
@@ -175,41 +177,41 @@ namespace Lolighter.Methods
                                 switch (n.CutDirection)
                                 {
                                     case CutDirection.Up:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Middle, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Middle, n.Type, CutDirection.Any));
                                         if (n.Type == NoteType.Red)
                                         {
-                                            newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.MiddleRight, Layer.Top, n.Type, CutDirection.Any));
+                                            newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.MiddleRight, Layer.Top, n.Type, CutDirection.Any));
                                         }
                                         else
                                         {
-                                            newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.Right, Layer.Top, n.Type, CutDirection.Any));
+                                            newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.Right, Layer.Top, n.Type, CutDirection.Any));
                                         }
                                         break;
                                     case CutDirection.Down:
                                         n.LineLayer = Layer.Top;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Middle, n.Type, CutDirection.Any));
-                                        newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.Right, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Middle, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.Right, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.Left:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleRight, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleRight, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.Right:
                                         n.LineIndex = Index.MiddleRight;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.UpLeft:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleLeft, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleLeft, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.UpRight:
                                         n.LineIndex = Index.MiddleRight;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Middle, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Middle, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.DownLeft:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleRight, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleRight, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.DownRight:
                                         n.LineIndex = Index.MiddleRight;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                 }
                                 break;
@@ -222,20 +224,20 @@ namespace Lolighter.Methods
                                 switch (n.CutDirection)
                                 {
                                     case CutDirection.Up:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.Down:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.UpRight:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleLeft, Layer.Top, n.Type, CutDirection.Any));
-                                        newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.MiddleRight, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleLeft, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.MiddleRight, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.DownLeft:
                                         break;
                                     case CutDirection.DownRight:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleLeft, Layer.Bottom, n.Type, CutDirection.Any));
-                                        newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.MiddleRight, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleLeft, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.MiddleRight, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                 }
                                 break;
@@ -265,18 +267,18 @@ namespace Lolighter.Methods
                                 switch (n.CutDirection)
                                 {
                                     case CutDirection.Up:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.Down:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.UpLeft:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleRight, Layer.Top, n.Type, CutDirection.Any));
-                                        newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.MiddleLeft, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleRight, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.MiddleLeft, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.DownLeft:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleRight, Layer.Bottom, n.Type, CutDirection.Any));
-                                        newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.MiddleLeft, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleRight, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.MiddleLeft, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                 }
                                 break;
@@ -290,40 +292,40 @@ namespace Lolighter.Methods
                                 {
                                     case CutDirection.Up:
                                         n.LineLayer = Layer.Bottom;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Middle, n.Type, CutDirection.Any));
-                                        newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.Left, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Middle, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.Left, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.Down:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Middle, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Middle, n.Type, CutDirection.Any));
                                         if (n.Type == NoteType.Red)
                                         {
-                                            newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.Left, Layer.Bottom, n.Type, CutDirection.Any));
+                                            newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.Left, Layer.Bottom, n.Type, CutDirection.Any));
                                         }
                                         else
                                         {
-                                            newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.MiddleLeft, Layer.Bottom, n.Type, CutDirection.Any));
+                                            newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.MiddleLeft, Layer.Bottom, n.Type, CutDirection.Any));
                                         }
                                         break;
                                     case CutDirection.Left:
                                         n.LineIndex = Index.MiddleLeft;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.Right:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleLeft, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleLeft, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.UpLeft:
                                         n.LineLayer = Layer.Middle;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.UpRight:
                                         n.LineLayer = Layer.Middle;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.DownLeft:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Middle, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Middle, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.DownRight:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleRight, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleRight, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                 }
                                 break;
@@ -331,27 +333,27 @@ namespace Lolighter.Methods
                                 switch (n.CutDirection)
                                 {
                                     case CutDirection.Down:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.MiddleLeft, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.MiddleLeft, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.Left:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.Right:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleRight, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleRight, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.UpLeft:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.UpRight:
                                         n.LineIndex = Index.Left;
                                         n.LineLayer = Layer.Middle;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleLeft, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleLeft, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.DownLeft:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Middle, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Middle, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.DownRight:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                 }
                                 break;
@@ -359,27 +361,27 @@ namespace Lolighter.Methods
                                 switch (n.CutDirection)
                                 {
                                     case CutDirection.Down:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.MiddleRight, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.MiddleRight, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.Left:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleLeft, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleLeft, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.Right:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.UpLeft:
                                         n.LineIndex = Index.Right;
                                         n.LineLayer = Layer.Middle;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleRight, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleRight, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.UpRight:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.DownLeft:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Left, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Left, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.DownRight:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Middle, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Middle, n.Type, CutDirection.Any));
                                         break;
                                 }
                                 break;
@@ -388,40 +390,40 @@ namespace Lolighter.Methods
                                 {
                                     case CutDirection.Up:
                                         n.LineLayer = Layer.Bottom;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Middle, n.Type, CutDirection.Any));
-                                        newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.Right, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Middle, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.Right, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.Down:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Middle, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Middle, n.Type, CutDirection.Any));
                                         if (n.Type == NoteType.Blue)
                                         {
-                                            newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.Right, Layer.Bottom, n.Type, CutDirection.Any));
+                                            newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.Right, Layer.Bottom, n.Type, CutDirection.Any));
                                         }
                                         else
                                         {
-                                            newNote.Add(new BeatmapNote(n.Time + 0.0625f, Index.MiddleRight, Layer.Bottom, n.Type, CutDirection.Any));
+                                            newNote.Add(new BeatmapNote(n.Time + thirdNote, Index.MiddleRight, Layer.Bottom, n.Type, CutDirection.Any));
                                         }
                                         break;
                                     case CutDirection.Left:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleRight, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleRight, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.Right:
                                         n.LineIndex = Index.MiddleRight;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.UpLeft:
                                         n.LineLayer = Layer.Middle;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.UpRight:
                                         n.LineLayer = Layer.Middle;
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Top, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Top, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.DownLeft:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.MiddleLeft, Layer.Bottom, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.MiddleLeft, Layer.Bottom, n.Type, CutDirection.Any));
                                         break;
                                     case CutDirection.DownRight:
-                                        newNote.Add(new BeatmapNote(n.Time + 0.03125f, Index.Right, Layer.Middle, n.Type, CutDirection.Any));
+                                        newNote.Add(new BeatmapNote(n.Time + secondNote, Index.Right, Layer.Middle, n.Type, CutDirection.Any));
                                         break;
                                 }
                                 break;